# docs/openapi/v1-iam-snapshot.yaml
openapi: 3.0.3
info:
  title: "cobo_iam_services — public & internal API snapshot"
  version: "1.0.0"
  description: |
    Danh sach endpoint dang dang ky trong `internal/httpserver` (handler modules). Mau JSON chi tiet: `api-contracts-json.md`. Bao gom: health, auth (login-password-key, login, refresh, logout, recovery, select/switch), me/*, sessions, admin, disclosure, workflow, notification, noi bo authorize.
  x-oas-source: api-v1-implemented-contracts.json
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
paths:
  "/api/v1/admin/companies/{company_id}/memberships":
    get:
      summary: "Liet ke member trong company"
      operationId: "get_api_v1_admin_companies__company_id__memberships"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships":
    post:
      summary: "Tao membership (admin)"
      operationId: "post_api_v1_admin_memberships"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}":
    patch:
      summary: "Cap nhat membership"
      operationId: "patch_api_v1_admin_memberships__membership_id_"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
    delete:
      summary: "Xoa membership"
      operationId: "delete_api_v1_admin_memberships__membership_id_"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}/departments":
    post:
      summary: "Gan department"
      operationId: "post_api_v1_admin_memberships__membership_id__departments"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}/departments/{department_id}":
    delete:
      summary: "Go department"
      operationId: "delete_api_v1_admin_memberships__membership_id__departments__department_id_"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}/roles":
    post:
      summary: "Gan role"
      operationId: "post_api_v1_admin_memberships__membership_id__roles"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}/roles/{role_id}":
    delete:
      summary: "Go role"
      operationId: "delete_api_v1_admin_memberships__membership_id__roles__role_id_"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}/titles":
    post:
      summary: "Gan title"
      operationId: "post_api_v1_admin_memberships__membership_id__titles"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/memberships/{membership_id}/titles/{title_id}":
    delete:
      summary: "Go title"
      operationId: "delete_api_v1_admin_memberships__membership_id__titles__title_id_"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/notification-rules":
    post:
      summary: "Tao quy tac thong bao"
      operationId: "post_api_v1_admin_notification_rules"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/permissions":
    get:
      summary: "Liet ke permission (catalog)"
      operationId: "get_api_v1_admin_permissions"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/resource-scope-rules":
    post:
      summary: "Tao quy tac scope tai nguyen"
      operationId: "post_api_v1_admin_resource_scope_rules"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/roles":
    get:
      summary: "Liet ke role theo company (policy)"
      operationId: "get_api_v1_admin_roles"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/roles/{role_id}/permissions":
    post:
      summary: "Gan quyen vao role"
      operationId: "post_api_v1_admin_roles__role_id__permissions"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/roles/{role_id}/permissions/{permission_id}":
    delete:
      summary: "Go quyen khoi role"
      operationId: "delete_api_v1_admin_roles__role_id__permissions__permission_id_"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/users":
    post:
      summary: "Tao user truc tiep (admin)"
      operationId: "post_api_v1_admin_users"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
  "/api/v1/admin/workflow-assignee-rules":
    post:
      summary: "Tao quy tac nguoi duoc giao buoc workflow"
      operationId: "post_api_v1_admin_workflow_assignee_rules"
      responses:
        "200": { description: OK }
        default:
          description: Error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
      security: [ { bearerAuth: [] } ]
...